fix: suppress empty mismatch warning for issue URLs; split helpers.go
- fix(suggest): CheckInputPipelineMismatch now returns nil when reason is
  empty, preventing a bare "  warning: " line from printing when input is
  an issue URL or free text (#774 follow-up)
- chore(cmd): split cmd/wave/commands/helpers.go into focused files:
  format_helpers.go (formatSize, formatTokens, truncateString, formatElapsed)
  and run_helpers.go (parseDuration, checkOnboarding) — no behavior change

Closes #1075
